guidea_engine/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.10
PyJWT==2.10.1
python-frontmatter==1.1.0
PyYAML==5.1
setuptools==80.9.0
sqlparse==0.5.3
